refactoring - splitting into more components
